first step refactoring: move mappers to jdbc package
